Uploaded new sidebar.js file.
Built a new sidebar.js with all the four new menus and all uploaded pages. There is probably some errors (and I don't know how to implement it within the design) so it will need someone to go through it, and then hook it up to the docosaurus system, and nav bar menu items.